% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phrdw_data_type.R
\name{assign_phrdw_data_type}
\alias{assign_phrdw_data_type}
\title{Assigns a data type to specific columns. This function will only assign a numeric datatype for fields ending
with _id or _key}
\usage{
assign_phrdw_data_type(phrdw_dataset, phrdw_datamart)
}
\arguments{
\item{phrdw_dataset}{A data frame with columns from the PHRDW.}
}
\value{
A data frame with numeric fields assigned.
}
\description{
Assigns a data type to specific columns. This function will only assign a numeric datatype for fields ending
with _id or _key
}
