% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_cd.R
\name{cd_investigation_query}
\alias{cd_investigation_query}
\title{prevents filtering non-exisitng columns and expandable by
allowing for additional user input that's not listed
converts dates to proper SQL format and removes
bad/non entries
this should cover all cases of incorrect
user input dates}
\usage{
cd_investigation_query(mart, type, param_list)
}
\description{
prevents filtering non-exisitng columns and expandable by
allowing for additional user input that's not listed
converts dates to proper SQL format and removes
bad/non entries
this should cover all cases of incorrect
user input dates
}
