% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_phrdw.R
\name{close_connection_to_phrdw}
\alias{close_connection_to_phrdw}
\title{Closes ODBC connection to PHRDW data marts.}
\usage{
close_connection_to_phrdw(phrdw_conn)
}
\arguments{
\item{phrdw_conn}{}
}
\description{
Closes ODBC connection to PHRDW data marts.
}
