% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_phrdw.R
\name{connect_to_phrdw}
\alias{connect_to_phrdw}
\title{Creates a connection object for PHRDW datamart based on user input}
\usage{
connect_to_phrdw(
  phrdw_datamart = NULL,
  mart = NULL,
  type = NULL,
  .conn_str = NULL,
  .return_conn_str = F
)
}
\arguments{
\item{phrdw_datamart}{`r lifecycle::badge('superseded')` Original mart
designations provided by previous package authors. This backward-
compatibility is meant to minimize changes on the user end.}

\item{mart}{Provide an appropriate mart name (non-case specific).
See `Details`.}

\item{type}{Provide an appropriate mart type (non-case specific).}

\item{.conn_str}{Defaults to `NULL`. For advance usage or testing purposes:
 if you are clear on the exact connection parameters,
 you can enter here as a named list, where name of
element is `cube` or `sql`, which will determine the appropriate connection
driver, and the element being the character string containing the specific
parameters.
See `Details`.}

\item{.return_conn_str}{}
}
\value{
An `odbc` or `OLAP_Conn` connection object that can be
executed with appropriate queries to retrieve views.
}
\description{
Connect to PHRDW data marts. Depending on the mart, the correct driver will
be selected and connection parameters will be populated automatically.
}
\details{
For a detailed list of data marts and respective servers, please see
`phrdwRdata::servers`.

For developers, connections can be expanded by expanding
`phrdwRdata::servers`.

List of data marts that can be connected to using `phrdw_datamart` and their
full name:
\itemize{
   \item CD: Communicable Diseases
   \item CDI: Chronic Disease & Injury
   \item PHRDW_Enteric_Panorama:
   \item SU_PHRDW_Enteric_Panorama:
   \item PHRDW_STIBBI:
   \item SU_PHRDW_STIBBI:
   \item SA_PHRDW_STIBBI:
   \item PHRDW_VPD:
   \item SU_PHRDW_VPD:
}

Alternatively, use a combination of the below:
List 0f marts:
\itemize{
   \item CD:
   \item CDI:
   \item Respiratory:
   \item Enteric:
   \item STIBBI:
   \item VPD:
}

list of mart types:
\itemize{
   \item prod: production
   \item su: UAT
   \item sa: Staging
}
}
\examples{
connect_to_phrdw('STIBBI')
connect_to_phrdw('Respiratory')
connect_to_phrdw(mart = 'STIBBI', type = 'prod')
}
