% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_batch_script.R
\name{create_r_batch_script}
\alias{create_r_batch_script}
\title{Generates Powershell commands for batch script.}
\usage{
create_r_batch_script(
  path_r_exe = NA_character_,
  path_proj_dir = NA_character_,
  path_run_script = NA_character_,
  path_std_log = NULL,
  .check_o_dirve = F,
  .set_pandoc = F,
  .load_renv = F,
  .renv_run = F,
  ...
)
}
\arguments{
\item{path_r_exe}{full executable path to Rscript.exe. If \code{NA}, will
automatically be determined based on the version of R in use.}

\item{path_proj_dir}{folder path to project. If \code{NA}, will automatically
invoke \code{getwd()}.}

\item{path_run_script}{user must supply the path to the script to execute.
Can supply without project directory for convenience. The function will
check file existence.}

\item{path_std_log}{full file path to a log, where standard outputs will be
redirected to. If \code{NULL}, a project_name.log will be created in a
"log" directory under project root. If no log should be produced, this can
accept boolean value \code{False}.}

\item{.check_o_dirve}{boolean value. Defaults to \code{False}. If \code{True}, adds
line to check if O drive is mapped, and will map if not.}

\item{.set_pandoc}{boolean value. Defaults to \code{False}. If \code{True}, adds
line to set pandoc location using \code{rmarkdown::find_pandoc()}.}

\item{.load_renv}{boolean value. Defaults to \code{False}. If \code{True}, adds
line to check if \code{renv} is installed, and will install if not. This will
also explicitly load project with \code{renv::load()}.}

\item{.renv_run}{boolean value. Defaults to \code{False}, and uses \code{source()}
to run script. If \code{True}, will use \code{renv::run()} to run script.}

\item{...}{users can supply further values to be assigned to variables.
The name of the parameter corresponds to the variable, and the value will
be assigned per Powershell syntax.}
}
\description{
Helps users create the appropriate Powershell commands per user
specifications, such as Rscript.exe location, pandoc, standard log outputs,
etc.
This function will parameterize by setting respective variables for easy
manual editing.
}
