% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype_cols.R
\name{datatype_cols}
\alias{datatype_cols}
\title{Cast data types appropriately.}
\usage{
datatype_cols(dataset, date_col_pattern = "date")
}
\arguments{
\item{dataset}{A \code{data.frame} or \code{tibble} object.}

\item{date_col_pattern}{String pattern to detect column names. Common
patterns can be 'date', or 'dt_tm'. Defaults to 'date'.}
}
\value{
A \code{data.frame} or \code{tibble} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Instead of heuristically prase dates, it detects date string by preset
patterns and apply respective \code{lubrdate} functions accordingly.
}
