% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute2D.R
\name{execute2D}
\alias{execute2D}
\title{Executes an OLAP Cube connection and a 2-dimention MDX query}
\usage{
execute2D(olapCnn, mdx_qry)
}
\arguments{
\item{olapCnn}{OLAP connection.}

\item{mdx_qry}{MDX query.}
}
\description{
Executes an OLAP Cube connection and a 2-dimention MDX query
}
