% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeMD.R
\name{executeMD}
\alias{executeMD}
\title{Executes an OLAP Cube connection and a multi-dimention MDX query}
\usage{
executeMD(olapCnn, mdx_qry)
}
\arguments{
\item{olapCnn}{OLAP connection.}

\item{mdx_qry}{MDX query.}
}
\description{
Executes an OLAP Cube connection and a multi-dimention MDX query
}
