% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Wrapper to explore cube names, dimensions, hierarchies, and levels of cube
connection.}
\usage{
explore(olapCnn, cube = NULL, dimension = NULL, hierarchy = NULL, level = NULL)
}
\arguments{
\item{level}{}
}
\description{
Wrapper to explore cube names, dimensions, hierarchies, and levels of cube
connection.
}
