% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phrdw_cdi_data.R
\name{get_phrdw_cdi_data}
\alias{get_phrdw_cdi_data}
\title{Get PHRDW CDI Data}
\usage{
get_phrdw_cdi_data(
  phrdw_datamart_connection,
  phrdw_datamart,
  dataset_name,
  query_start_date,
  query_end_date,
  ucd_3char_code = "",
  ccd_3char_code = "",
  residential_location_ha = "",
  death_location_ha = ""
)
}
\arguments{
\item{phrdw_datamart_connection}{A PHRDW connection object.}

\item{phrdw_datamart}{The data mart to query.}

\item{dataset_name}{The dataset to retrieve.}

\item{query_start_date}{The start date of the data to retrieve.}

\item{query_end_date}{The end date of the data to retrieve.}

\item{ucd_3char_code}{A character vector of UCD Codes to retrieve. Optional parameter.}

\item{ccd_3char_code}{A character vector of CCD Codes to retrieve. Optional parameter.}

\item{residential_location_ha}{A character vector of BC Health Authorities associated with decedent's usual residence. Optional parameter.}

\item{death_location_ha}{A character vector of BC Health Authorities associated with the location of death. Optional parameter.}
}
\value{
A data frame with the dataset retrieved from the specified PHRDW data mart.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Function for querying and returning PHRDW CDI data
}
