% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phrdw_data.R
\name{get_phrdw_data}
\alias{get_phrdw_data}
\title{Title}
\usage{
get_phrdw_data(
  phrdw_datamart_connection = NULL,
  phrdw_datamart = NULL,
  dataset_name = NULL,
  query_start_date = NULL,
  query_end_date = NULL,
  include_patient_identifiers = F,
  include_indigenous_identifiers = F,
  retrieve_system_ids = "Yes",
  disease = NULL,
  surveillance_condition = NULL,
  classification = NULL,
  surveillance_region_ha = NULL,
  infection_group = NULL,
  ordering_provider_ha = NULL,
  lis_status = NULL,
  episode_status = NULL,
  test_type = NULL,
  episode_testing_pattern = NULL,
  testing_region_ha = NULL,
  case_status = NULL,
  case_source = NULL,
  collect_data = T,
  mart = NULL,
  type = NULL,
  ...
)
}
\arguments{
\item{phrdw_datamart_connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Original
function design: supply a connection object created by \link{connect_to_phrdw}.
Recommend using \code{mart} and \code{type} instead for flexibility
(see \link{connect_to_phrdw}). The function takes care of connecting
to the appropriate PHRDW database and disconnect after performing the
requested data filtering and retrieving.}

\item{phrdw_datamart}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Original function
design: supply the \emph{exact} name of the datamart. Need to exercise case
sensitivity. Recommend using \code{mart} and \code{type} instead for flexibility
(see \link{connect_to_phrdw}).}

\item{mart}{Provide an appropriate mart name (non-case specific).
See \code{Details}.}

\item{type}{Provide an appropriate mart type (non-case specific).
See \code{Details}.}

\item{...}{}
}
\description{
List of marts:
\itemize{
\item CD:
\item CDI:
\item Respiratory:
\item Enteric:
\item STIBBI:
\item VPD:
}
}
\details{
list of mart types:
\itemize{
\item prod: production
\item su: UAT
\item sa: Staging
}
}
