% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phrdw_data_legacy.R
\name{get_phrdw_data_legacy}
\alias{get_phrdw_data_legacy}
\title{Get PHRDW Data}
\usage{
get_phrdw_data_legacy(
  phrdw_datamart_connection,
  phrdw_datamart,
  dataset_name,
  query_start_date,
  query_end_date,
  include_patient_identifiers = FALSE,
  include_indigenous_identifiers = FALSE,
  retrieve_system_ids = "Yes",
  disease = "",
  surveillance_condition = "",
  classification = "",
  surveillance_region_ha = "",
  infection_group = "",
  ordering_provider_ha = "",
  lis_status = "",
  episode_status = "",
  test_type = "",
  episode_testing_pattern = "",
  testing_region_ha = "",
  case_status = "",
  case_source = ""
)
}
\arguments{
\item{phrdw_datamart_connection}{A PHRDW connection object.}

\item{phrdw_datamart}{The data mart to query.}

\item{dataset_name}{The dataset to retrieve.}

\item{query_start_date}{The start date of the data to retrieve.}

\item{query_end_date}{The end date of the data to retrieve.}

\item{include_patient_identifiers}{Sets whether to retrieve patient identifiers in a query. Only applicable to CD Data Mart. Default is FALSE.}

\item{include_indigenous_identifiers}{Sets whether to retrieve patient identifiers in a query. Only applicable to CD Data Mart. Default is FALSE.}

\item{retrieve_system_ids}{Sets whether to retrieve system ids in a query. Currenly Only applicable to Enteric Data Mart. Default is "Yes".}

\item{disease}{A character vector of diseases to retrieve. Optional parameter.}

\item{surveillance_condition}{A character vector of surveillance conditions to retrieve. Optional parameter.}

\item{classification}{A character vector of classifications to retrieve. Optional parameter.}

\item{surveillance_region_ha}{A character vector of Health Region Authorities to retrieve. Use this variable to pull data by where the Patient lived. Optional parameter.}

\item{infection_group}{A character vector of infection groups to retrieve. Optional parameter.}

\item{ordering_provider_ha}{A character vector of Health Region Authorities to retrieve. Use this variable to pull data by Ordering Provider Health Authorities. Optional parameter.}

\item{lis_status}{A character vector of case level statuses about the LIS data in a Case. Optional parameter.}

\item{episode_status}{A character vector of episode statuses from the LIS result processing and rule engine. Optional parameter.}

\item{test_type}{A character vector of the types of tests to retrieve. Optional parameter.}

\item{episode_testing_pattern}{A character vector of testing patterns to retrieve. Optional parameter.}

\item{testing_region_ha}{A character vector of the testing regions to retrieve. Optional parameter.}

\item{case_status}{A character vector of the case statuses to retrieve. Optional parameter.}

\item{case_source}{A character vector of the case sources to retrieve. Optional parameter.}
}
\value{
A data frame with the dataset retrieved from the specified PHRDW data mart.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Function for querying and returning PHRDW data
}
