% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sql_view.R
\name{map_sql_view}
\alias{map_sql_view}
\title{Retrieve full mapping of views and columns using \emph{sys.schema}.}
\usage{
map_sql_view(
  conn = NULL,
  catalog = NULL,
  schema = NULL,
  tbl_vw_dependencies = F
)
}
\arguments{
\item{tbl_vw_dependencies}{}
}
\description{
Retrieve full mapping of views and columns using \emph{sys.schema}.
}
