% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx_builder.R
\name{mdx_build}
\alias{mdx_build}
\title{MDX builder: takes all MDX functions and build query.}
\usage{
mdx_build(
  cube_name,
  columns,
  rows,
  dim_props,
  .partial,
  discrete = NULL,
  range = NULL
)
}
\arguments{
\item{cube_name}{Cube name.}

\item{columns}{Character vector of measures, with name of dimension as the
name of the list. Name defaults to \code{Measures}.}

\item{rows}{Accepts \code{data.frame} with columns \code{dim}, \code{attr_hier}, and
\code{lvl_memb}, or character list of hierarchies, with name of dimension as
the name of the list.}

\item{dim_props}{Must be \code{data.frame} with columns \code{dim}, \code{attr_hier}, and
\code{lvl_memb}.}

\item{.partial}{A named list of a single element, where name is either
\code{head} or \code{tail}, and element is an integer. This will modify query to
retrieve the first or last parts of data frame going by the default
order.}

\item{discrete}{A \code{data.frame} object with 3 columns: \code{dim}, \code{attr}, and
\code{memb}, for "dimension", "attribute", and "member". The attribute must
belong to dimension, and member must belong to the attribute hierarchy.
Each member to filter for should have its own row in this \code{data.frame}.}

\item{range}{A \code{data.frame} object with 3 columns: \code{dim}, \code{attr}, and
\code{memb}, for "dimension", "attribute", and "member".
The attribute must belong to dimension, and member must belong to the
attribute hierarchy.
Two rows must be provided here with two different member values as
the \code{from} and \code{to}.

If no bounds, use \code{NULL} or "null".}
}
\value{
A \code{sql}/\code{character} object.
}
\description{
MDX builder: takes all MDX functions and build query.
}
\examples{
\dontrun{
phrdwRdata::mdx_build(
  cube_name = 'StibbiDM',
  columns = 'Case Count',
  rows =
    set_names(
      list(
        c(
          "Age Group 10",
          "Age Group 24",
          "Age Years"
        )
      ),
      'Case - Age at Earliest Date'
    ),
  discrete =
    tibble(
      dim = 'LIS - Test',
      attr = 'Test Code',
      memb = c('TPE1', 'RPR')
    ),
  range =
    tibble(
      dim = 'LIS - Date - Collection',
      attr = 'Date',
      memb = c('2019-01-01', '2019-02-02')
    )
)
}

}
