% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx_builder.R
\name{mdx_from}
\alias{mdx_from}
\title{MDX Builder: builds the \code{FROM} clause, and incorporate any filters if needed.}
\usage{
mdx_from(cube_name, ...)
}
\arguments{
\item{cube_name}{Cube name.}

\item{...}{Character vector of lines, which makes up the filter query
build from \code{mdx_filters}.}
}
\value{
A \code{sql}/\code{character} object.
}
\description{
MDX Builder: builds the \code{FROM} clause, and incorporate any filters if needed.
}
\details{
mdx_from(
'StibbiDM',
mdx_filter(
discrete =
tibble(
dim = 'LIS - Test',
attr = 'Test Code',
memb = c('TPE1', 'RPR')
),
range =
tibble(
dim = 'LIS - Date - Collection',
attr = 'Date',
memb = c('2019-01-01', '2019-02-02')
),
.as_lines = T
)
)
}
