% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olap_handler.R
\name{olap_handler}
\alias{olap_handler}
\title{Helper function for handling and building MDX queries.}
\usage{
olap_handler()
}
\value{
Depending on user input, \code{tibble}, MDX string, or print dataset
metadata.

In user_params and default_params, there can be some user
inputs for filtering. This selects for them, match them up to
respective dim-hierarchy by hierarchy names or arg names,
rearranges into data frame of columns \code{dim} for dimension,
\code{attr} for attribute hierarchy, and \code{memb} for the user input
part to select for, formatted for mdx_builder.
}
\description{
Takes arguments from \code{\link[=get_phrdw_data]{get_phrdw_data()}} and processes with
\code{phrdwRdata:::list_query_info$olap} to produce MDX query accordingly
}
