% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlapConnection.R
\name{OlapConnection}
\alias{OlapConnection}
\title{Constructs an "OlapConnection" object.}
\usage{
OlapConnection(conn_str = "Data Source=localhost; Provider=MSOLAP;")
}
\arguments{
\item{conn_str}{Connection string that requires \verb{Initial Catalog}, which is
the name of the data mart, \verb{Data Source}, \code{Provider=MSOLAP}, and
\verb{Packet Size=32767} for PHRDW OLAP server.}
}
\value{
A character string with class \code{OLAP_Conn}.
}
\description{
string of connection parameters that can be executed with MDX queries
using C routines for efficiency.
Checks if connection string contains all the necessary components to connect
to OLAP.
}
