% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrdwRdata-package.R
\docType{package}
\name{phrdwRdata-package}
\alias{phrdwRdata}
\alias{phrdwRdata-package}
\title{phrdwRdata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikuo0628.github.io/phrdwRdata/}
}

}
\keyword{internal}
