% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlapConnection.R
\name{print.OlapConnection}
\alias{print.OlapConnection}
\title{Simple helper function to print \code{OLAP_conn} object.}
\usage{
\method{print}{OlapConnection}(mdx_qry)
}
\arguments{
\item{mdx_qry}{}
}
\description{
Simple helper function to print \code{OLAP_conn} object.
}
