% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{Query}
\alias{Query}
\title{Construct OLAP query.}
\usage{
Query(validate = F)
}
\arguments{
\item{validate}{}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Query` constructs a "Query" object. Set functions are used to build and
modify the Query axes and cube name.
}
