% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{Renames data retrieved from Data cubes.}
\usage{
rename_cols(obj)
}
\arguments{
\item{obj}{A \code{data.frame} or character vector of names.}
}
\value{
Same as input `obj.
}
\description{
Renames data retrieved from Data cubes.
}
