% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_phrdw_columns.R
\name{rename_phrdw_columns}
\alias{rename_phrdw_columns}
\title{Rename PHRDW Columns}
\usage{
rename_phrdw_columns(phrdw_dataset)
}
\arguments{
\item{phrdw_dataset}{A data frame with columns from the PHRDW.}
}
\value{
A data frame with renamed columns from a PHRDW query
}
\description{
Function renames the columns for queries executed against a PHRDW Datamart
}
