% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_handler.R
\name{sql_handler}
\alias{sql_handler}
\title{Helper function for handling and building SQL queries.}
\usage{
sql_handler()
}
\value{
Depending on user input, \code{tibble}, SQL string, or print dataset
metadata.
}
\description{
Takes arguments from \code{\link[=get_phrdw_data]{get_phrdw_data()}} and processes with
\code{phrdwRdata:::list_query_info$sql} leverging \code{dbplyr} and NSE to produce
SQL query accordingly.
}
