% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Wrapper to explore cube names, dimensions, hierarchies, and levels of cube
connection.}
\usage{
explore(olapCnn, cube = NULL, dimension = NULL, hierarchy = NULL, level = NULL)
}
\arguments{
\item{olapCnn}{Required. An OLAP connection string.
Can be created with \code{connect_to_phrdw}.}

\item{cube}{Optional. Cube name.}

\item{dimension}{Optional. Dimension name.}

\item{hierarchy}{Optional. Hierarchy name.}

\item{level}{Optional. Level name.}
}
\value{
Character vector of cube(s) information.
}
\description{
Given \code{olapCnn} OLAP connection string, returns info on cubes under the OLAP
server. If no cube information is supplied, it will return all the
information of the highest level.
}
