% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{invisible_Call}
\alias{invisible_Call}
\title{Wrapper to capture print output from calling C/C++ OLAP routines.}
\usage{
invisible_Call(...)
}
\arguments{
\item{...}{}
}
\description{
Wrapper to capture print output from calling C/C++ OLAP routines.
}
