% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx_builder.R
\name{mdx_select}
\alias{mdx_select}
\title{MDX Builder: builds the SELECT statement on \code{columns} and \code{rows}.}
\usage{
mdx_select(columns, rows)
}
\arguments{
\item{columns}{Character vector of measures, with name of dimension as the
name of the list. Name defaults to \code{Measures}.}

\item{rows}{Character list of hierarchies, with name of dimension as the
name of the list.}
}
\value{
A \code{sql}/\code{character} object.
}
\description{
MDX Builder: builds the SELECT statement on \code{columns} and \code{rows}.
}
\details{
mdx_select(
'Case Count',
set_names(
list(
c(
"Age Group",
"Age Group 04",
"Age Group 05",
"Age Group 09",
"Age Group 10",
"Age Group 11",
"Age Group 17",
"Age Group 20",
"Age Group 24",
"Age Years"
)
),
'Case - Age at Earliest Date'
)
)
}
