% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx_builder.R
\name{mdx_select}
\alias{mdx_select}
\title{MDX Builder: builds the SELECT statement on \code{columns} and \code{rows}.}
\usage{
mdx_select(columns, rows, dim_props, .partial = NULL)
}
\arguments{
\item{columns}{Character vector of measures, with name of dimension as the
name of the list. Name defaults to \code{Measures}.}

\item{rows}{Accepts \code{data.frame} with columns \code{dim}, \code{attr_hier}, and
\code{lvl_memb}, or character list of hierarchies, with name of dimension as
the name of the list.}

\item{dim_props}{Must be \code{data.frame} with columns \code{dim}, \code{attr_hier}, and
\code{lvl_memb}.}

\item{.partial}{A named list of a single element, where name is either
\code{head} or \code{tail}, and element is an integer. This will modify query to
retrieve the first or last parts of data frame going by the default
order.}
}
\value{
A \code{sql}/\code{character} object.
}
\description{
MDX Builder: builds the SELECT statement on \code{columns} and \code{rows}.
}
\examples{
\dontrun{
mdx_select(
  'Case Count',
  set_names(
    list(
      c(
        "Age Group",
        "Age Group 04",
        "Age Group 05",
        "Age Group 09",
        "Age Group 10",
        "Age Group 11",
        "Age Group 17",
        "Age Group 20",
        "Age Group 24",
        "Age Years"
      )
    ),
    'Case - Age at Earliest Date'
  )
)
}

}
