% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olap_handler.R
\name{olap_handler}
\alias{olap_handler}
\title{Helper function to process dataset instructions into MDX query.}
\usage{
olap_handler()
}
\value{
MDX query.

In user_params and default_params, there can be some user
inputs for filtering. This selects for them, match them up to
respective dim-hierarchy by hierarchy names or arg names,
rearranges into data frame of columns \code{dim} for dimension,
\code{attr} for attribute hierarchy, and \code{memb} for the user input
part to select for, formatted for mdx_builder.
}
\description{
Takes arguments in \code{get_phrdw_data} and returns MDX query to be executed
in \code{execute2D}.
}
