% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrdwRdata-package.R
\docType{package}
\name{phrdwRdata-package}
\alias{phrdwRdata}
\alias{phrdwRdata-package}
\title{phrdwRdata: Retrieve pre-built PHSA Datasets}
\description{
Efficient interface for for retrieving pre-built datasets. An ease-of-use wrapper that allows robust user defined filtering while abstracting away backend database connections, queries, complex joins, and dataset specific conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikuo0628.github.io/phrdwRdata/}
}

}
\author{
\strong{Maintainer}: Michael Kuo \email{michael.kuo@bccdc.ca}

Authors:
\itemize{
  \item Darren Frizzell (Original Author)
}

Other contributors:
\itemize{
  \item  BCCDC [copyright holder, funder]
}

}
\keyword{internal}
