% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{Query}
\alias{Query}
\title{Construct OLAP query.
\code{Query} constructs a "Query" object. Set functions are used to build and
modify the Query axes and cube name.}
\usage{
Query(validate = F)
}
\arguments{
\item{validate}{}
}
\description{
Construct OLAP query.
\code{Query} constructs a "Query" object. Set functions are used to build and
modify the Query axes and cube name.
}
