% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_handler.R
\name{sql_handler}
\alias{sql_handler}
\title{default dplyr/dbplyr joins do not keep keys (for good reasons),
and some keys are needed/SELECT.
the \code{keep} param in join functions will determine whether
keys are kept. For keys that are part of SELECT: T... this is
done by checking to see if it's needed in \code{SELECT} clause or
if it does not already exists in df_1}
\usage{
sql_handler()
}
\description{
default dplyr/dbplyr joins do not keep keys (for good reasons),
and some keys are needed/SELECT.
the \code{keep} param in join functions will determine whether
keys are kept. For keys that are part of SELECT: T... this is
done by checking to see if it's needed in \code{SELECT} clause or
if it does not already exists in df_1
}
